Add files via upload
The Statistics and EDA project aims to provide training and testing on fundamental Data Exploratory and 
Statistical techniques used in the industry. It covers descriptive and inferential methods, as well as in-depth 
cleaning and analysis of a dataset to extract valuable business insights. This project will familiarize you with 
the common practices of data scientists, particularly in the area of Exploratory Data Analysis. 

Key Skills: Exploratory Data Analysis, Analyzing using Statistical Models